Bug 1502115 - Add @Annotation support to apilint. r=snorp.
This patch adds support for @Annotations in the API specification.

apilint will track certain, hard-coded, annotations and will error out if any
of those annotations are added or removed as that is assumed to be an
incompatible change.

This is the initial list of annotations that are tracked is as following:

- java.lang.Deprecated
- android.support.annotation.AnyThread
- android.support.annotation.BinderThread
- android.support.annotation.MainThread
- android.support.annotation.UiThread
- android.support.annotation.WorkerThread